+++ include/qpdf/DLL.h
@@ -27,12 +27,12 @@
 /* The first version of qpdf to include the version constants is 10.6.0. */
 #define QPDF_MAJOR_VERSION 12
 #define QPDF_MINOR_VERSION 0
-#define QPDF_PATCH_VERSION 0
+#define QPDF_PATCH_VERSION 1
 
 #ifdef QPDF_FUTURE
-# define QPDF_VERSION "12.0.0+future"
+# define QPDF_VERSION "12.0.1+future"
 #else
-# define QPDF_VERSION "12.0.0"
+# define QPDF_VERSION "12.0.1"
 #endif
 
 /*